+++ modules/pytket-braket/pytket/extensions/braket/backends/braket.py
@@ -107,6 +107,8 @@ OQC_SCHEMA = {
 }
 
 _gate_types = {
+    "amplitude_damping": None,
+    "bit_flip": None,
     "ccnot": OpType.CCX,
     "cnot": OpType.CX,
     "cphaseshift": OpType.CU1,
@@ -117,12 +119,18 @@ _gate_types = {
     "cv": OpType.CV,
     "cy": OpType.CY,
     "cz": OpType.CZ,
+    "depolarizing": None,
     "ecr": OpType.ECR,
     "end_verbatim_box": None,
+    "generalized_amplitude_damping": None,
     "h": OpType.H,
     "i": OpType.noop,
     "iswap": OpType.ISWAPMax,
+    "kraus": None,
+    "pauli_channel": None,
     "pswap": None,
+    "phase_damping": None,
+    "phase_flip": None,
     "phaseshift": OpType.U1,
     "rx": OpType.Rx,
     "ry": OpType.Ry,
@@ -133,6 +141,8 @@ _gate_types = {
     "swap": OpType.SWAP,
     "t": OpType.T,
     "ti": OpType.Tdg,
+    "two_qubit_dephasing": None,
+    "two_qubit_depolarizing": None,
     "unitary": None,
     "v": OpType.V,
     "vi": OpType.Vdg,
@@ -160,6 +170,8 @@ _multiq_gate_types = {
     "iswap",
     "pswap",
     "swap",
+    "two_qubit_dephasing",
+    "two_qubit_depolarizing",
     "unitary",
     "xx",
     "xy",
@@ -190,14 +202,27 @@ def _obs_from_qpo(operator: QubitPauliOperator, n_qubits: int) -> Observable:
 
 def _get_result(
     completed_task: Union[AwsQuantumTask, LocalQuantumTask],
+    n_qubits: int,
     want_state: bool,
+    want_dm: bool,
     ppcirc: Optional[Circuit] = None,
 ) -> Dict[str, BackendResult]:
     result = completed_task.result()
     kwargs = {}
-    if want_state:
-        kwargs["state"] = result.get_value_by_result_type(ResultType.StateVector())
+    if want_state or want_dm:
         assert ppcirc is None
+        if want_state:
+            kwargs["state"] = result.get_value_by_result_type(ResultType.StateVector())
+        if want_dm:
+            m = result.get_value_by_result_type(
+                ResultType.DensityMatrix(target=list(range(n_qubits)))
+            )
+            if type(completed_task) == AwsQuantumTask:
+                kwargs["density_matrix"] = np.array(
+                    [[complex(x, y) for x, y in row] for row in m], dtype=complex
+                )
+            else:
+                kwargs["density_matrix"] = m
     else:
         kwargs["shots"] = OutcomeArray.from_readouts(result.measurements)
         kwargs["ppcirc"] = ppcirc
@@ -332,6 +357,12 @@ class BraketBackend(Backend):
             elif rtname == "Variance":
                 self._variance_min_shots = rtminshots
                 self._variance_max_shots = rtmaxshots
+            elif rtname == "DensityMatrix":
+                self._supports_density_matrix = True
+                # Always use n_shots = 0 for DensityMatrix
+        # Don't use contextual optimization for non-QPU backends
+        if self._device_type != _DeviceType.QPU:
+            self._supports_contextual_optimisation = False
 
         self._singleqs, self._multiqs = self._get_gate_set(
             supported_ops, self._device_type
@@ -572,8 +603,9 @@ class BraketBackend(Backend):
 
     @property
     def _result_id_type(self) -> _ResultIdTuple:
-        # (task ID, whether state vector is wanted, serialized ppcirc or "null")
-        return (str, bool, str)
+        # (task ID, whether state vector / density matrix are wanted, serialized ppcirc
+        # or "null")
+        return (str, int, bool, bool, str)
 
     def _run(
         self, bkcirc: braket.circuits.Circuit, n_shots: int = 0, **kwargs: KwargTypes
@@ -627,7 +659,8 @@ class BraketBackend(Backend):
 
         handles = []
         for circ, n_shots in zip(circuits, n_shots_list):
-            want_state = n_shots == 0
+            want_state = (n_shots == 0) and self.supports_state
+            want_dm = (n_shots == 0) and self.supports_density_matrix
             if postprocess:
                 circ_measured = circ.copy()
                 circ_measured.measure_all()
@@ -638,6 +671,8 @@ class BraketBackend(Backend):
             bkcirc = self._to_bkcirc(c0)
             if want_state:
                 bkcirc.add_result_type(ResultType.StateVector())
+            if want_dm:
+                bkcirc.add_result_type(ResultType.DensityMatrix(target=bkcirc.qubits))
             if not bkcirc.instructions and len(circ.bits) == 0:
                 task = None
             else:
@@ -646,16 +681,26 @@ class BraketBackend(Backend):
                 # Results are available now. Put them in the cache.
                 if task is not None:
                     assert task.state() == "COMPLETED"
-                    results = _get_result(task, want_state, ppcirc)
+                    results = _get_result(
+                        task, bkcirc.qubit_count, want_state, want_dm, ppcirc
+                    )
                 else:
                     results = {"result": self.empty_result(circ, n_shots=n_shots)}
             else:
                 # Task is asynchronous. Must wait for results.
                 results = {}
             if task is not None:
-                handle = ResultHandle(task.id, want_state, json.dumps(ppcirc_rep))
+                handle = ResultHandle(
+                    task.id,
+                    bkcirc.qubit_count,
+                    want_state,
+                    want_dm,
+                    json.dumps(ppcirc_rep),
+                )
             else:
-                handle = ResultHandle(str(uuid4()), False, json.dumps(None))
+                handle = ResultHandle(
+                    str(uuid4()), bkcirc.qubit_count, False, False, json.dumps(None)
+                )
             self._cache[handle] = results
             handles.append(handle)
         return handles
@@ -671,7 +716,7 @@ class BraketBackend(Backend):
     def circuit_status(self, handle: ResultHandle) -> CircuitStatus:
         if self._device_type == _DeviceType.LOCAL:
             return CircuitStatus(StatusEnum.COMPLETED)
-        task_id, want_state, ppcirc_str = handle
+        task_id, n_qubits, want_state, want_dm, ppcirc_str = handle
         ppcirc_rep = json.loads(ppcirc_str)
         ppcirc = Circuit.from_dict(ppcirc_rep) if ppcirc_rep is not None else None
         task = AwsQuantumTask(task_id, aws_session=self._aws_session)
@@ -681,7 +726,9 @@ class BraketBackend(Backend):
         elif state == "CANCELLED":
             return CircuitStatus(StatusEnum.CANCELLED)
         elif state == "COMPLETED":
-            self._update_cache_result(handle, _get_result(task, want_state, ppcirc))
+            self._update_cache_result(
+                handle, _get_result(task, n_qubits, want_state, want_dm, ppcirc)
+            )
             return CircuitStatus(StatusEnum.COMPLETED)
         elif state == "QUEUED" or state == "CREATED":
             return CircuitStatus(StatusEnum.QUEUED)
@@ -692,10 +739,15 @@ class BraketBackend(Backend):
 
     @property
     def characterisation(self) -> Optional[Dict[str, Any]]:
+        node_errors = self._backend_info.all_node_gate_errors
+        edge_errors = self._backend_info.all_edge_gate_errors
+        readout_errors = self._backend_info.all_readout_errors
+        if node_errors is None and edge_errors is None and readout_errors is None:
+            return None
         return {
-            "NodeErrors": self._backend_info.all_node_gate_errors,
-            "EdgeErrors": self._backend_info.all_edge_gate_errors,
-            "ReadoutErrors": self._backend_info.all_readout_errors,
+            "NodeErrors": node_errors,
+            "EdgeErrors": edge_errors,
+            "ReadoutErrors": readout_errors,
         }
 
     @property