+++ modules/pytket-braket/docs/changelog.rst
@@ -8,6 +8,7 @@ Changelog
 * Update requirements for amazon-braket-sdk and amazon-braket-schemas.
 * Add support for oqc provider.
 * Correct readout errors reported for Rigetti devices.
+* Add support for density-matrix simulators.
 
 0.16.0 (February 2022)
 ----------------------